+++ include/asm-arm/arch-ixp2000/io.h
@@ -16,7 +16,7 @@
 #define __ASM_ARM_ARCH_IO_H
 
 #define IO_SPACE_LIMIT		0xffffffff
-#define __mem_pci(a)		((unsigned long)(a))
+#define __mem_pci(a)		(a)
 
 /*
  * Pick up VMALLOC_END
@@ -24,7 +24,7 @@
 #define ___io(p)		((unsigned long)((p)+IXP2000_PCI_IO_VIRT_BASE))
 
 /*
- * IXP200 does not do proper byte-lane conversion for PCI addresses,
+ * IXP2000 does not do proper byte-lane conversion for PCI addresses,
  * so we need to override standard functions.
  */
 #define alignb(addr)		((addr & ~3) + (3 - (addr & 3)))
